[CLI] Remove PlaygroundCliWorker base class, inline php command logic
Remove the PlaygroundCliWorker intermediary class and restore the
original type union. The runCLIScript method is replaced with inline
stream handling that properly awaits pipeTo() promises via Promise.all,
eliminating the setTimeout workaround for output flushing.
