+++ packages/playground/cli/src/blueprints-v2/worker-thread-v2.ts
@@ -52,7 +52,7 @@ import type {
 } from '@wp-playground/wordpress';
 import { shouldRenderProgress } from '../utils/progress';
 import type { Mount } from '@php-wasm/cli-util';
-import { PlaygroundCliWorker } from '../playground-cli-worker';
+import { PHPWorker } from '@php-wasm/universal';
 
 async function mountResources(php: PHP, mounts: Mount[]) {
 	for (const mount of mounts) {
@@ -185,7 +185,7 @@ export type WorkerBootRequestHandlerOptions = Omit<
 	spawnHandler: () => SpawnHandler;
 };
 
-export class PlaygroundCliBlueprintV2Worker extends PlaygroundCliWorker {
+export class PlaygroundCliBlueprintV2Worker extends PHPWorker {
 	booted = false;
 	blueprintTargetResolved = false;
 	phpInstancesThatNeedMountsAfterTargetResolved = new Set<PHP>();
@@ -517,8 +517,13 @@ export class PlaygroundCliBlueprintV2Worker extends PlaygroundCliWorker {
 		}
 	}
 
-	override async mountAfterWordPressInstall(mounts: Array<Mount>) {
+	async mountAfterWordPressInstall(mounts: Array<Mount>) {
 		await mountResources(this.__internal_getPHP()!, mounts);
+	}
+
+	// Provide a named disposal method that can be invoked via comlink.
+	async dispose() {
+		await this[Symbol.asyncDispose]();
 	}
 
 	async applyPostInstallMountsToAllWorkers(
